#1 refactor: refactor code
