+++ src/components/Cube.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { useLoader } from "@react-three/fiber";
 import { TextureLoader } from "three/src/loaders/TextureLoader";
-import { colors } from "../../consts";
+import { colors } from "../consts";
 
 const Cube = ({ position, index, cubesRefHandler }) => {
